+++ src/migrations/m221017_120000_sync_campaign_reports.php
@@ -3,7 +3,7 @@
 namespace putyourlightson\campaign\migrations;
 
 use craft\db\Migration;
-use putyourlightson\campaign\records\CampaignRecord;
+use putyourlightson\campaign\Campaign;
 use putyourlightson\campaign\records\ContactCampaignRecord;
 
 class m221017_120000_sync_campaign_reports extends Migration
@@ -13,18 +13,7 @@ class m221017_120000_sync_campaign_reports extends Migration
      */
     public function safeUp(): bool
     {
-        /** @var CampaignRecord[] $campaignRecords */
-        $campaignRecords = CampaignRecord::find()
-            ->where(['>', 'recipients', '0'])
-            ->all();
-
-        foreach ($campaignRecords as $campaignRecord) {
-            $campaignRecord->opened = $this->_getCampaignColumnCount($campaignRecord->id, 'opened');
-            $campaignRecord->clicked = $this->_getCampaignColumnCount($campaignRecord->id, 'clicked');
-            $campaignRecord->opens = $this->_getCampaignColumnSum($campaignRecord->id, 'opens');
-            $campaignRecord->clicks = $this->_getCampaignColumnSum($campaignRecord->id, 'clicks');
-            $campaignRecord->save();
-        }
+        Campaign::$plugin->reports->sync();
 
         return true;
     }
@@ -37,20 +26,5 @@ class m221017_120000_sync_campaign_reports extends Migration
         echo self::class . " cannot be reverted.\n";
 
         return false;
-    }
-
-    private function _getCampaignColumnCount(int $campaignId, string $column): int
-    {
-        return (int)ContactCampaignRecord::find()
-            ->where(['campaignId' => $campaignId])
-            ->andWhere(['not', [$column => null]])
-            ->count();
-    }
-
-    private function _getCampaignColumnSum(int $campaignId, string $column): int
-    {
-        return (int)ContactCampaignRecord::find()
-            ->where(['campaignId' => $campaignId])
-            ->sum($column);
     }
 }